repaired solutions for final tasks

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,7 +3,7 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
 
-        int farenheitTemperature = (temperatureCelsius * 9) / 5 + 32;
+        float farenheitTemperature = (temperatureCelsius * 9) / 5f + 32;
         System.out.println(farenheitTemperature);
 
     }
